+++ LexerBase.pas
@@ -26,6 +26,7 @@ type
 procedure HandleCRLF(Parser: TParser; Formatter: TFormatterBase);
 procedure HandleSpace(Parser: TParser; Formatter: TFormatterBase);
 procedure HandleSlashesComment(Parser: TParser; Formatter: TFormatterBase);
+procedure HandleLineComment(Parser: TParser; Formatter: TFormatterBase; CommentMark: string);
 
 implementation
 
@@ -100,7 +101,12 @@ end;
 
 procedure HandleSlashesComment(Parser: TParser; Formatter: TFormatterBase);
 begin
-  if (Parser.Current = '/') and (Parser.PeekNext = '/') then
+  HandleLineComment(Parser, Formatter, '//');
+end;
+
+procedure HandleLineComment(Parser: TParser; Formatter: TFormatterBase; CommentMark: string);
+begin
+  if Parser.PeekLength(Length(CommentMark)) = CommentMark then
   begin
     while (not Parser.IsEof) and (not Parser.IsEoln) do
       Parser.Next;
@@ -110,4 +116,3 @@ begin
 end;
 
 end.
-